fixed more builds
