+++ applications/minotari_ledger_wallet/comms/src/ledger_wallet.rs
@@ -20,24 +20,14 @@
 // WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use std::{
-    fmt,
-    fmt::{Display, Formatter},
-    ops::Deref,
-};
+use std::ops::Deref;
 
 use ledger_transport::{APDUAnswer, APDUCommand};
 use ledger_transport_hid::{hidapi::HidApi, TransportNativeHID};
 use num_derive::FromPrimitive;
 use num_traits::FromPrimitive;
-use serde::{Deserialize, Serialize};
-use tari_common::configuration::Network;
-use tari_crypto::ristretto::RistrettoPublicKey;
 
 use crate::error::LedgerDeviceError;
-
-const WALLET_CLA: u8 = 0x80;
-// const LOG_TARGET: &str = "wallet::ledger_wallet::comms";
 
 #[repr(u8)]
 #[derive(FromPrimitive, Debug, Copy, Clone, PartialEq)]
@@ -72,6 +62,10 @@ pub struct Command<D> {
 }
 
 impl<D: Deref<Target = [u8]>> Command<D> {
+    pub fn new(inner: APDUCommand<D>) -> Command<D> {
+        Self { inner }
+    }
+
     pub fn execute(&self) -> Result<APDUAnswer<Vec<u8>>, LedgerDeviceError> {
         get_transport()?
             .exchange(&self.inner)
@@ -85,82 +79,5 @@ impl<D: Deref<Target = [u8]>> Command<D> {
         transport
             .exchange(&self.inner)
             .map_err(|e| LedgerDeviceError::NativeTransport(e.to_string()))
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LedgerWallet {
-    account: u64,
-    pub pubkey: Option<RistrettoPublicKey>,
-    pub network: Network,
-}
-
-impl Display for LedgerWallet {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        write!(f, "account {}", self.account)?;
-        write!(f, "pubkey {}", self.pubkey.is_some())?;
-        Ok(())
-    }
-}
-
-impl LedgerWallet {
-    pub fn new(account: u64, network: Network, pubkey: Option<RistrettoPublicKey>) -> Self {
-        Self {
-            account,
-            pubkey,
-            network,
-        }
-    }
-
-    pub fn account_bytes(&self) -> Vec<u8> {
-        self.account.to_le_bytes().to_vec()
-    }
-
-    pub fn build_command(&self, instruction: Instruction, data: Vec<u8>) -> Command<Vec<u8>> {
-        let mut base_data = self.account_bytes();
-        base_data.extend_from_slice(&data);
-
-        Command {
-            inner: APDUCommand {
-                cla: WALLET_CLA,
-                ins: instruction.as_byte(),
-                p1: 0x00,
-                p2: 0x00,
-                data: base_data,
-            },
-        }
-    }
-
-    pub fn chunk_command(&self, instruction: Instruction, data: Vec<Vec<u8>>) -> Vec<Command<Vec<u8>>> {
-        let num_chunks = data.len();
-        let mut more;
-        let mut commands = vec![];
-
-        for (i, chunk) in data.iter().enumerate() {
-            if i + 1 == num_chunks {
-                more = 0;
-            } else {
-                more = 1;
-            }
-
-            // Prepend the account on the first payload
-            let mut base_data = vec![];
-            if i == 0 {
-                base_data.extend_from_slice(&self.account_bytes());
-            }
-            base_data.extend_from_slice(chunk);
-
-            commands.push(Command {
-                inner: APDUCommand {
-                    cla: WALLET_CLA,
-                    ins: instruction.as_byte(),
-                    p1: u8::try_from(i).unwrap_or(0),
-                    p2: more,
-                    data: base_data,
-                },
-            });
-        }
-
-        commands
     }
 }
